refactor : SkillMapper 디렉토리 변경
