+++ src/main/java/univ/kgu/carely/domain/common/embeded/skill/util/SkillMapper.java
@@ -1,6 +1,7 @@
-package univ.kgu.carely.domain.common.embeded.skill;
+package univ.kgu.carely.domain.common.embeded.skill.util;
 
 import org.mapstruct.Mapper;
+import univ.kgu.carely.domain.common.embeded.skill.Skill;
 import univ.kgu.carely.domain.member.dto.request.ReqUpdateSkillDTO;
 
 @Mapper(componentModel = "spring")
